feat: add business rules and rents page

+++ src/pages/Cars/index.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable no-restricted-globals */
 import React, { useEffect, useState } from 'react';
 import Modal from 'react-modal';
 
@@ -11,17 +12,16 @@ import api from '../../services/api';
 
 import styles from './styles.module.scss';
 
-interface CarProps {
+export interface CarProps {
   id?: number;
   name: string;
   brand: string;
-
   year: string;
-  docNumber: string;
-  chassi: string;
-  licensePlate: string;
   category: string;
-  client: string;
+  docNumber: string;
+  licensePlate: string;
+  timesRented?: number;
+  status: "RENTING" | "NOT_RENTING";
 }
 
 interface CategoryProps {
@@ -41,8 +41,9 @@ export function Cars() {
   const [carYear, setCarYear] = useState('');   
   const [carDocNumber, setCarDocNumber] = useState('');   
   const [carLicensePlate, setCarLicensePlate] = useState('');   
-  const [carCategory, setCarCategory] = useState("Sedan");     
+  const [carCategory, setCarCategory] = useState('');     
   const [carBrand, setCarBrand] = useState('');   
+  const [carStatus, setCarStatus] = useState<string>("NOT_RENTING");
 
   const [carCategorys, setCarCategorys] = useState<CategoryProps[]>([]);
 
@@ -65,9 +66,10 @@ export function Cars() {
     setCarModel('');
     setCarBrand('');
     setCarYear('');  
-    setCarCategory("Sedan");
+    setCarCategory('');
     setCarDocNumber('');    
-    setCarLicensePlate('');    
+    setCarLicensePlate('');   
+    setCarStatus("NOT_RENTING");
   }
 
   async function getCars() {
@@ -95,6 +97,8 @@ export function Cars() {
         docNumber: carDocNumber,
         licensePlate: carLicensePlate,
         category: carCategory,
+        status: carStatus,
+        timesRented: 0,
       });
     } else {
       alert("Preencha os campos corretamente!");           
@@ -104,9 +108,10 @@ export function Cars() {
     setCarModel('');
     setCarBrand('');
     setCarYear('');  
-    setCarCategory("Sedan");
+    setCarCategory('');
     setCarDocNumber('');    
-    setCarLicensePlate('');    
+    setCarLicensePlate('');
+    setCarStatus("NOT_RENTING");   
     
     getCars();
   }
@@ -115,14 +120,15 @@ export function Cars() {
     event.preventDefault();
     closeModalEdit();
     
-    if(carModel && carBrand && carYear && carDocNumber && carLicensePlate && carCategory) {
-      await api.put(`/cars/${carID}`, {             
+    if(carModel && carBrand && carYear && carDocNumber && carLicensePlate && carCategory && carStatus) {
+      await api.patch(`/cars/${carID}`, {             
         name: carModel,
         brand: carBrand,
         year: carYear,
         docNumber: carDocNumber,
         licensePlate: carLicensePlate,
-        category: carCategory,
+        category: carCategory,    
+        status: carStatus    
       });
     } else {
       alert("Preencha os campos corretamente!");           
@@ -132,29 +138,46 @@ export function Cars() {
     setCarModel('');
     setCarBrand('');
     setCarYear('');  
-    setCarCategory("Sedan");
+    setCarCategory('');
     setCarDocNumber('');    
-    setCarLicensePlate('');    
+    setCarLicensePlate(''); 
+    setCarStatus("NOT_RENTING");   
     
     getCars();
   }
 
   async function DeleteCar(id?: number) {
-    // eslint-disable-next-line no-restricted-globals
-    const confirmation = confirm("Deletar carro?");  
+    if(cars![id! - 1].status === "RENTING") {
+      alert("Não é possível deletar o carro pois existe um aluguel ativo com este carro");
+    } else {
+      const confirmation = confirm("Deletar carro?");  
     
-    if(confirmation) {    
-      await api.delete(`/cars/${id}`);
-      getCars();
+      if(confirmation) {  
+        const reason = prompt("Qual a razão da exclusão do carro?");
+
+        const car = await api.get(`/cars/${id}`);
+
+        await api.post('/car-deleted/', {
+          car: car.data,
+          reason: reason
+        });
+        
+        await api.delete(`/cars/${id}`);
+        getCars();
+      }
     }
   }
 
-  async function HandleEditCar(id?: number) {        
-    await api.get(`/cars/${id}`).then((response) => {
-      setCar(response.data);       
-      console.log(response.data);    
-    })    
-    openModalEdit();
+  async function HandleEditCar(id?: number) {       
+    if(cars![id! - 1].status === "RENTING") {
+      alert("Não é possível editar o carro pois existe um aluguel ativo com este carro");
+    } else { 
+      await api.get(`/cars/${id}`).then((response) => {
+        setCar(response.data);       
+        console.log(response.data);    
+      })    
+      openModalEdit();
+    }
   }
   
   useEffect(() => {
@@ -162,11 +185,11 @@ export function Cars() {
       setCarID(car.id);
       setCarModel(car.name);
       setCarBrand(car.brand);
-
       setCarYear(car.year);  
       setCarCategory(car.category);
       setCarDocNumber(car.docNumber);    
       setCarLicensePlate(car.licensePlate);      
+      setCarStatus(car.status);
     }    
   }, [car]);    
 
@@ -205,14 +228,9 @@ export function Cars() {
 
             <p>Categoria:</p>
             <select 
-              onChange={e => {
-                if(e.target.value.length > 1) {
-                  setCarCategory(e.target.value);
-                } else {
-                  setCarCategory("Sedan");
-                }              
-              }}
+              onChange={e => setCarCategory(e.target.value)}              
             >
+              <option hidden>Selecione uma opção</option>
               {carCategorys.map((category, index) => {                
                 return (
                   <option key={index}>{category.name}</option>
